% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Determine the probability 2 or more people share the same birthday in a group of x number of people.}
\usage{
birthday(x)
}
\arguments{
\item{x}{Size of group of people}
}
\value{
Percentage of Probability
}
\description{
Determine the probability 2 or more people share the same birthday in a group of x number of people.
}
\examples{
birthday(20)
}
