% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireDamData.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage and Distance Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from the fire station (in miles)}
  \item{DAMAGE}{Fire damage (in thousands of dollars)}
}
}
\source{
[MS 6th]
}
\usage{
fire
}
\description{
A dataset containing measurements of fire damage in relation to distance
from the fire station. This data can be used to examine the relationship
between response distance and the extent of fire damage.
}
\keyword{datasets}
