% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQuadCurve.R
\name{fitQuadCurve}
\alias{fitQuadCurve}
\title{Quadratic Curve Fitter}
\usage{
fitQuadCurve(lm, ylim, xlim, dataframe)
}
\arguments{
\item{lm}{a linear model with quadratic vector}

\item{ylim}{The value to scale y}

\item{xlim}{The value to scale x}

\item{dataframe}{A dataframe that corresponds to the model}
}
\value{
scatter plot
}
\description{
Fit a Quadratic Curve to its data frame scatterplot
}
\examples{
data(cars)
car.lm <- lm(dist ~ speed + I(speed^2), data = cars)
fitQuadCurve(car.lm, c(0, 150), c(0, 30), cars)
}
