% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCltP.R
\name{mycltp}
\alias{mycltp}
\title{Creates a Poisson distribution across a number of iteration of size n to show how approximations become more accurate and n increases.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{size of samples}

\item{iter}{number of sample}

\item{lambda}{lambda to use for the Poission distribution}

\item{...}{Any additional parameters to override}
}
\value{
A 1,1,2,3 Plot of the Sampling mean, the rel freq of y values, and the probability distribution of y
}
\description{
Creates a Poisson distribution across a number of iteration of size n to show how approximations become more accurate and n increases.
}
\examples{
mycltp(n=2,iter=10000, lambda = 4)
}
