% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{The number of trials}

\item{N}{Sample size to draw from}

\item{r}{Number of success elements in N}

\item{n}{Number of elements to draw}
}
\value{
iter number of percentage distributions for each count of drawing and a bar chart
}
\description{
Takes the arguments to make a distribution of drawing successes among a number of trials
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
