% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(mu = 1, sigma = 2, a = 1)
}
\arguments{
\item{mu}{the mean of the curve (peak)}

\item{sigma}{the average variation about the curve}

\item{a}{the probability to shade calculate up until a}
}
\value{
Normal Curve with a shaded area
}
\description{
My Normal Curve
}
\examples{
myncurve(1,2,1)
}
