% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice1.R
\name{myNewFunc}
\alias{myNewFunc}
\title{Plot the x^2 curve for a range}
\usage{
myNewFunc(x)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The y = x^.
}
\description{
Plot the x^2 curve for a range
}
\examples{
myNewFunc(1:10)
}
