% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter plot with marginal histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{quantitative vector}

\item{y}{quantitative vector}

\item{xlab}{x label string}

\item{ylab}{y label string}
}
\value{
a scatter plot
}
\description{
Scatter plot with marginal histogram
}
\examples{
scatterhist(x = 1:10, y = 11:20, xlab = "x", ylab = "y")
}
